+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * Added `DESC` user field to store user description (e.g. user biography) for public display
     * Added code highlight to components
     * Added `get_page_field()` theme function to get value from requested field for current page
+    * Added `get_page_author()` theme function to get value from `author` field for current page
 
 1. [](#improved)
     * Updated theme function `get_navigation()` to add `aria-role="page"` to current page in menu